Fix for 32-bit
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@5790 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/re/re_parser.h
@@ -150,7 +150,7 @@ protected:
     inline void require(std::string s) {
         if (!accept(s)) {
             if (mCursor.noMore()) ParseFailure("Expecting " + s + " but end of input encountered");
-            unsigned long rem = mCursor.remaining();
+            size_t rem = mCursor.remaining();
             ParseFailure("Expecting " + s + " but " + std::string(mCursor.pos(), mCursor.pos() + std::min(rem, s.length())) + " encountered");
         }
     }
